gsdx-hw: Clamp zbuffer to depthmask instead of wrapping.
The behavior was verified on Dobie to be correct.
The code needs to be ported to SW renderer too to
improve rendering on SW side.

Current PR will fix plenty of games on HW renderer
that zbuffer issues before.

v2. Set DepthMask to the maximum the current depth format allows.
Will properly clamp for 16 and 24bit formats.